fixed translate to reduce to int - double translations lead to tiny drawing errors in case of non-full clip
git-svn-id: https://svn.code.sf.net/p/genj/code/trunk@1146 fbd980a2-0dff-11df-b4b7-0b05b74d0985

+++ genj/src/core/genj/util/swing/UnitGraphics.java
@@ -89,7 +89,7 @@ public class UnitGraphics {
   /**
    * Translation   */
   public void translate(double dx, double dy) {
-    graphics.translate(dx * unitx, dy * unity);
+    graphics.translate((int)(dx * unitx), (int)(dy * unity));
   }
   
   /**
@@ -174,6 +174,7 @@ public class UnitGraphics {
    * Draw shape
    */
   public void draw(Shape shape, double x, double y, boolean fill) {
+    
     AffineTransform at = new AffineTransform();
     at.scale(unitx, unity);
     at.translate(x,y);
